move build step into separate job
This separates out the steps to build the tarball used for deployments
into its own job, making it no longer dependent on the unit tests
completing before the updated tarball can be deployed according to the
rules defined in the repository's `SOURCE_TARGET_MAP` JSON variable.
This will allow deployments to be made without waiting for 2 full rounds
of unit tests to complete, as is currently the case.

However, the current branch protection rules do not ensure that the unit
tests required by a PR were run on the same code that will be deployed
if the PR is merged. This is because the unit tests are run on the head
branch of the PR, while the deployment is made from the merged result of
the PR. When the 2 full tests were being run, this was not an issue
because the 2nd test run was done on the merged result of the PR, and
the deploy job would wait for that 2nd test run to succeed.

To recoup the protection of the 2nd test run while keeping the speed
gained by breaking out these steps, another branch protection rule must
be added to the repo requiring the head branch of a PR to be fully
updated with the base branch of the PR. This will ensure all tests which
were run on the head branch will correctly reflect how they will run on
the merged result of the PR. This will also ensure that the deployment
will be made from the same code that was tested.

The only negative to adding this rule is that it will require all PRs
merging changes up the chain of `develop` to `staging` to `master` to be
done using proxy branches, since those branches cannot be updated with
the base branch of the PR directly due to branch protection rules.

This also performs a few cleanup items on the workflow file and related
scripts.