change invoice_schedule join to left join (#443)
Unit tests are failing because the inner join is causing all invoice
queries to return 0 rows. This will allow them so succeed while the
schedules table is still empty.

This is a temporary fix til we get a proper update in place for the
invoices data loader to always create a payment schedule
when an invoice is created.